add screens in one folder
