+++ app/signupSignin/SignUp.tsx
@@ -1,16 +1,16 @@
 import React, { useState } from "react";
 import { View, StyleSheet, Image, Text } from "react-native";
 
-import { containerStyle, titleStyle } from "../styles/common";
-import Button from "../components/button";
-import SignUpText from "../components/SignUpText";
-import BackIcon from "../components/BackIcon";
-import EmailPasswordInput from "../components/EmailPasswordInput";
-import RememberMe from "../components/RememberMe";
-import SignUpWith from "../components/SignUpWith";
-import Or from "../components/Or";
+import { containerStyle, titleStyle } from "../../styles/common";
+import Button from "../../components/button";
+import SignUpText from "../../components/SignUpText";
+import BackIcon from "../../components/BackIcon";
+import EmailPasswordInput from "../../components/EmailPasswordInput";
+import RememberMe from "../../components/RememberMe";
+import SignUpWith from "../../components/SignUpWith";
+import Or from "../../components/Or";
 
-const SingIn = () => {
+const SingUp = () => {
   const [isEmailActive, setIsEmailActive] = useState<boolean>(false);
   const [isPasswordActive, setIsPasswordActive] = useState<boolean>(false);
 
@@ -25,7 +25,7 @@ const SingIn = () => {
   return (
     <View style={{ ...containerStyle, rowGap: 24 }}>
       <BackIcon imageSource={require("../assets/Account.png")} />
-      <Text style={titleStyle}>Login to Your Account</Text>
+      <Text style={titleStyle}>Create New Account</Text>
       <View style={{ rowGap: 20 }}>
         <EmailPasswordInput
           icon="email"
@@ -41,16 +41,16 @@ const SingIn = () => {
       </View>
       <RememberMe />
       <Button
-        title="Sign in"
+        title="Sign up"
         backgroundColor={
           isEmailActive || isPasswordActive ? "#1E90FF" : "#3062C8"
         }
       />
       <Or text="or continue with" />
       <SignUpWith />
-      <SignUpText text1="Don’t have an account?" text2="Sign up" />
+      <SignUpText text1="Already have an account?" text2="Sign in" />
     </View>
   );
 };
 
-export default SingIn;
+export default SingUp;
